Refactored iterator rules
